Changes from expat-2.2.0@17579827 rebased onto 2.4.1
These changes have removed SKIP_PRE_BUILD_COMMAND from CMake and defines in build scripts. It conflicted in the merge as the upstream cmake script is heavily modified. This option seems to have designed to prevent execution of documentation generation prior to the lib compile. In 2.4.1 documentation generation has it's own conditional, so we don't need our own. Documentation generation is also hidden behind the tools conditional, which we already disable.

#rnx
#jira VLN-1330
#rb trivial

[CL 17820650 by geoff evans in 5.0 branch]