DUMMY MERGE (Accept Target) GPU hang workaround.
Add define COMPILER_HINT_REDUCE_REGISTER_PRESSURE to allow shaders to hint to the compiler to reduce register pressure at the expense of occupancy. Values range from 1 to 4

Enable this for VirtualTextureMaterial in order to workaround a GPU hang on one platform where register spillage can result in bad fetches due to a shader compiler bug. This may actually be faster in the general case, regardless of the GPU hang.

We will probably revisit this with a more targetted solution when we have one


#rb jamie.hayes

#ROBOMERGE-OWNER: Ben.Woodhouse
#ROBOMERGE-AUTHOR: ben.woodhouse
#ROBOMERGE-SOURCE: CL 17843089 via CL 17843094 via CL 17843101
#ROBOMERGE-BOT: (v883-17842818)
#ROBOMERGE-CONFLICT from-shelf

[CL 17843180 by Ben Woodhouse in 5.0 branch]