Fix issues with connecting to the correct trace session.
The TraceDataFiltering and TraceSourceFiltering plugins were designed to work inside the editor with one live session. When communicating the newest live session was used. This was problematic when used in Unreal Insights and more than one live session was active. With this change the correct address and port for the current live session is used.

#rb ionut.matasaru
#jira UE-101954
#preflight 616e94887f15cb0001c5fdd2

[CL 17858578 by Johan Berg in 5.0 branch]