SlateDebugging - Adding a lot more metadata options by including the actual event, which we now use when printing the data, assuming it has any to print, like for mouse input events it can now tell you where the cursor is which is fantastic.  Adding a few new slate events to the log, to help debug problems with cursor muckery by higher UI systems.
Slate Application - When attempting to return to using the OS Cursor, from a state of using the Faux Slate cursor for gamepad input, we now require at least 3 events before we transition back.  This is because the OS is continuing to send us mouse input, for non-mouse movements.  I believe it to be related to either mouse capture stopping, or transioning away from high precision mouse movement disabling, or potentially cursor lock.  Anyway - it's some non-easily determinable source, but the solution is still the same, we need to avoid a handful of events before we trust the OS signaling that we really need to go back to using the OS cursor due to an honest mouse move.

#review-17883975 @Patrick.Boutot

[CL 17883973 by Nick Darnell in 5.0 branch]