Fix issue with auto import sample rate. We dont want to use the blendshape or material curve in the logic to find the best bake sample rate since those curve are not baked.
Fix the curve import when the curve is weighted. We need to provide the hypotenuse length instead of the raw normalize weight to richcurve to have the proper result of weighting information.

The fbx sequencer curve import is now using the same code has the fbx content asset import.

Add a test to test fbx curve with weighted tangents import. I also fix a path issue with the fbx automation test tool.

#jira UE-130597
#rb michael.zyracki
#preflight 6176f7f7487b42000115d6b9

[CL 17911258 by Alexis Matte in 5.0 branch]