Added support for removing categories from sections.
This is specifically to cover the case where too many categories end up in the same section, reducing the usefulness of each individual section. One way to get this to occur is to have multiple EditInline UObjects in the same class. This functionality allows a programmer to trim down the list to what they want it to be.

Added tests for said functionality.

[at]editor-ux


#ROBOMERGE-AUTHOR: sebastian.nordgren
#ROBOMERGE-SOURCE: CL 18032045 via CL 18032080 via CL 18032100
#ROBOMERGE-BOT: (v885-17909292)

[CL 18032475 by sebastian nordgren in 5.0 branch]