UE5 + Instanced Stereo Rendering: Skip check for re-used LoadBalancers in ShadowProjection pass
FProjectedShadowInfo::RenderProjection will re-use the ProjectionStencilingPass for the left and right eye. When we go through the right-eye path when ISR is enabled, we re-use the contained FInstanceCullingContext. FInstanceCullingContext::SetupDrawCommands doesn't expect to be re-used, and fails a check where the LoadBalancers are empty.

I can skip the check if ISR is enabled, and the results seem correct.

Future-looking: support ISR natively on shadow projections

#jira UE-123704
#rb ola.olsson jules.blok steve.smith

[CL 18141418 by robert srinivasiah in 5.0 branch]