Removed all assumptions that FText must have, or be able to have, a shared display string
This assumption persisted because the original FText implementation included a shared display string pointer, which even generated texts would create. Later (~2015) the ITextData pointer replaced the shared display string pointer, which allowed generated texts to avoid allocating the shared display string (unless serialized or rebuilt from a culture change).

The downside of this was the that ITextData implementations were a complex extra layer of indirection, and generated texts still held a display string pointer even though it was null 99% of the time.

This change removes that assumption, and removes the complexity of the ITextData implementations by instead having each FTextHistory control how it wants to manage and rebuild its display string, whether that be caching a shared display string from the text localization manager or generating it on-demand.

Note: ITextData still exists, but it is now directly implemented by FTextHistory. This is slightly redundant from a C++ sense, but allows natvis to function correctly since FTextHistory is a private type.

Changes:
 - Removed FTextHistory::SerializeForDisplayString. This data is now handled directly by FTextHistory::Serialize, with FTextHistory::UpdateDisplayString resolving the display string when needed.
 - Added FTextHistory_Generated to provide the common functionality for generated text histories that need to re-build a local display string.
 - Added FTextSnapshot::LocalizedStringPtr to detect if the localized string changed within a history.
   - This handles FTextHistory_StringTableEntry lazily resolving its display string (it used to reset its history version to force the update, but that's no longer viable as it would result in an infinite loop of UpdateDisplayString).
 - FTextHistory_StringTableEntry now caches the localized display string, rather than FStringTableEntry, so that it can handle updating it when needed.
 - Updated FStringTable to use FTextKey rather than FString for text namespace and keys.
 - FTextLocalizationManager::FindDisplayString and FTextLocalizationManager::GetDisplayString now return immutable shared display strings, in preparation for changing FTextLocalizationManager to share display strings.

#jira
#rb Danny.Couture
#preflight 61926eac9f7a6462e0069329


#ROBOMERGE-AUTHOR: jamie.dale
#ROBOMERGE-SOURCE: CL 18190731 via CL 18190965 via CL 18191210
#ROBOMERGE-BOT: (v895-18170469)

[CL 18192780 by jamie dale in 5.0 branch]