Allow FTextLocalizationManager to preserve the de-duplicated display string data loaded from a LocRes
Now that text no longer assumes that display strings are unique, or that every text must have a mutable shared display string instance, it's pretty trivial to allow FTextLocalizationManager/FTextLocalizationResource to preserve the de-duplicated display string data loaded from a LocRes file.

Note: FTextLocalizationManager itself does not de-duplicate display strings (as that would require another LUT, which would cost more than it saves). Instead it relies on the fact that the LocRes data is already de-duplicated, and just preserves that de-duplicated state on load. While that won't be perfect (as identical display strings in different LocRes won't be shared) it should still catch 98-99% of duplicated text... for free!

 - Made FDisplayStringEntry::DisplayString const, as shared display strings cannot be mutated.
 - Updated FTextLocalizationResource to preserve its de-duplicated display string data, to be applied within FTextLocalizationManager.
 - Leetification (for debug) now makes unique display string instances (as it cannot modify the shared display string).
 - Removed FDisplayStringEntry::bIsLocalized as it was inconsistently set, and ultimately not significantly used.

#jira
#rb Danny.Couture
#preflight 6192a8b6c80d0ce51a7e0fa0


#ROBOMERGE-AUTHOR: jamie.dale
#ROBOMERGE-SOURCE: CL 18195297 via CL 18195498 via CL 18195537
#ROBOMERGE-BOT: (v895-18170469)

[CL 18200475 by jamie dale in 5.0 branch]