Sequencer: Control Rig: Add ability to edit multiple controls
Instead of one details panel with topmost nodes we have a detail panel per type, and then allow for multiple objects for each.  When theres multiple objects the system still asks for each customization for each object so we need to store the 'multiple' name so the title is correct. Also FTransform would not work correctly with multiple types so instead we are using FEulerTransform with coercion. (Note FTransforms aren't common anymore FEulerTransform is really the default CR transform type but we have lots of legacy assets still using Transform controls)

#jira UE-135517
#preflight 6196ebbab22f9e02b57d6211
#rb max.chen

[CL 18248485 by Mike Zyracki in 5.0 branch]