Switchboard: Command line now puts -game before map name.
This avoids an issue where if the map name coincides with a commandlet, the commandlet will be run. In particular, SmokeTest map was triggering the run of SmokeTestCommandlet.

By having -game before the map, it triggers a slightly different initialization path that doesn't have this behavior. It still finds the correct map name because -game is moved to the end of the command line, so the map name becomes the new first token.

#rb
#jira UE-136719

[CL 18392884 by Alejandro Arango in 5.0 branch]