SocialAnalytics, convert social user analytics context to TMap to allow arbitrary number of contextual analytics params
When we added the SocialMenuContext analytic we added the ability to execute a social interaction "with context". At the time the context was implemented as a string. We now have a request for more functionality here, so converting the string to a dictionary to allow for more params.

#tests
launched client with -AnalyticsDisableCaching unfriended and re-friended someone, checked the analytics string in the logs for the proper values



#ROBOMERGE-AUTHOR: austin.ellis
#ROBOMERGE-SOURCE: CL 18538675 via CL 18538682 via CL 18538695 via CL 18538701
#ROBOMERGE-BOT: (v899-18417669)

[CL 18539382 by austin ellis in 5.0 branch]