MovieScene: query movie scenes and sequences for their earliest TimecodeSource instead of using their own property
UMovieScene is a container that may represent multiple pieces of media as
sections, each of which may have their own distinct timecode source. As a result,
the timecode source associated with the UMovieScene itself doesn't correspond to
any particular piece of media, but rather was just intended as the timecode at which
the movie scene was recorded.

These changes update code sites that read the TimecodeSource property directly to
instead query the UMovieScene/UMovieSceneSequence for the earliest timecode source
of any of its sections using GetEarliestTimecodeSource(). The TimecodeSource property
of UMovieScene will ultimately be removed in a subsequent change.

#rb max.chen
#preflight 61dcc052e67256ec41f23837

[CL 18586470 by matt johnson in 5.0 branch]