MovieScene: remove TimecodeSource property stored directly on UMovieScene
These changes remove the last remnants of the TimecodeSource property that
was being stored directly on UMovieScene. UMovieScene is a container that may
represent multiple pieces of media as UMovieSceneSections, each of which may
have their own distinct TimecodeSource. The sections themselves manage setting
their own TimecodeSource when they are recorded or created from a piece of media.

Client code that is interested in the earliest TimecodeSource of any section in a movie
scene should use UMovieScene::GetEarliestTimecodeSource().

#rb max.chen
#preflight 61e9bd06ab6e1cc028b89f96

[CL 18681887 by matt johnson in 5.0 branch]