Menus are not main windows (Mac)
I've added a boolean AllowMainWindow to CocoaWindow which is now set to true unless the window is a menu.  This status is reported to the OS via the canBecomeMainWindow method.  It is also used by the MacApplication which will now not activate a window when a window that's being closed is not a main window (via AllowMainWindow).

#rb will.damon
#jira UE-137460
#preflight n/a

[CL 18734334 by dave hunter in 5.0 branch]