MovieSceneTools: look for timecode properties for use as TimecodeSource when importing FBX to Control Rig
These changes add handling for custom timecode properties coming from FBX in
MovieSceneToolHelpers::ImportFBXIntoControlRigChannels(). All nodes being imported
from FBX are searched for animated properties that match the timecode custom attribute
names (which default to "TCHour", "TCMinute", "TCSecond", and "TCFrame", but can be
customized in the animation settings). When any of those properties are found on a node,
an FTimecode is constructed from the component values and is used to populate the
"TimecodeSource" property of all of the movie scene sections that were modified by the
import operation. The first node encountered that has timecode properties is the one
whose values will be used.

#rb max.chen, mike.zyracki
#preflight 61f0872efd5285142b285f4d

[CL 18738040 by matt johnson in 5.0 branch]