Fix incorrect geometric normals on path traced geometry with WPO materials
The post-process which applies WPO usually leaves mesh vertices in world space. However the vertex factory might still contains a local to world transform which is required to rotate the shading normals correctly. We thus end up with a mismatch when the world position is computed for ray tracing. Currently only the path tracer was manifesting an artifact due to this, but the inconsistency could have shown up in other ways (for example if a material had a dependency on world space position).

The fix is to simply keep the mesh in "local" space by instructing the RayTracingDynamicMesh compute shader to put the vertex back into local space after applying WPO.

All the geometry types were reviewed for this potential issue and the RayTracingDynamicMesh has been updated to mandate a preprocessor signal on how to handle this.

#jira UE-139634
#rb Yuriy.ODonnell, Juan.Canada
#preflight 61f871aa114ec25fe0bfe9b3

[CL 18801676 by chris kulla in 5.0 branch]